fix GQL snapshot tests on py310 (#11945)
### Summary & Motivation

A previous PR I merged had some `check` calls that failed only on 3.10
(related to checking a `Protocol`). This is causing BK master to fail.
This PR removes the check calls and fixes the error.

### How I Tested These Changes

BK with `test-py310`, ensuring tests failing on master are run in this
PR's feature branch.