[dagit] Support range-based partition health data (#11573)
### Summary & Motivation

This PR makes Dagit compile again with the updates in #10822

This is phase 1 of two phases of work to use range-based partition
health data. This PR updates our existing client-side health structures
to read the new GraphQL endpoints which return the state of partition
ranges rather than the state of each individual partition in a "matrix"
of booleans.

Mapping the new APIs to our client-side requirements does require some
transforms, and it increases code complexity in
`usePartitionHealthData`. Namely, we need to add the start/end partition
indexes of each range so we can efficiently look up the status of a
particular partition within the range data. (eg: `given these ranges of
[A->C], [E->F], is D materialized?`)

The second phase of work is to 1) make the "partition health bar" rely
more directly on this data and not load the partition key space along
with it, and 2) make Dagit query for the health of individual partitions
individually via a new endpoint (so it doesn't have to answer the
question above). This is our eventual goal, but will require a bit more
work:

- In the static partitions case, `materializedPartitions` alone cannot
tell us the correct display order for `materializedPartitions` /
`unmaterializedParttions`. We still need to load all the keys in order
in order to display the health bar.

- In the multi-dimensional health case, the returned structure contains
subranges. We don't actually care at all about the subranges, but we
need to know whether the second dimension is `"complete", "partial" or
"missing"`. I think a backend resolver might need to be added for this
that just returns these three states, or missingCount / completeCount.

We're planning to postpone the second chunk of work here, because the
current implementation (which loads the partition keys) isn't actually
that slow now that we've made other changes, and it's unclear if all the
work will be worthwhile.

Co-authored-by: bengotow <bgotow@elementl.com>