[pyright] [examples/docs-snippets] misc (#11921)
### Summary & Motivation

Fix some type errors (and a few straight-up runtime errors) in docs
snippets. Note that `# type: ignore` comments added to snippets are
stripped from their rendered form in the docs (I added this
functionality in a PR a few weeks ago). Notes:

- I fixed one error by adding an `assert isinstance` on some data loaded
from yaml, this seems to me to be the most universally understandable
approach for user-facing content.
- The `@schedule` correction in the airflow tutorial was approved by
@Ramshackle-Jamathon over slack

### How I Tested These Changes

BK