Refactor ProcessRegistryEntry (#11959)
### Summary & Motivation

This is a light refactor of `ProcessRegistryEntry` (a class that is used
in a very limited scope) with the goal of fixing type errors and
enhancing readability. It drops the `check` calls previously used in
construction, which I think is OK as discussed for internal objects in
this discussion: in this discussion:
https://github.com/dagster-io/internal/discussions/4618#discussioncomment-4760254

### How I Tested These Changes

Existing BK tests.