feat(dbt-cloud): compile run only if job has environment variable cache (#12042)
### Summary & Motivation
Some user feedback on the caching mechanism:
> Will this compile all jobs within that project ID - isn’t that a bit
unnecessary since that means all of my staging, testing jobs will be
compiled as well? Or will it reference my dagster repo code for only the
jobs necessary to compile?

> we have jobs just hanging around from tests (branch tests for complex
changes) , others that are ad hoc jobs (for people to manually update
and run), jobs just for when I test dbt upgrades, etc etc
:slightly_smiling_face: None of those need to be run

> All jobs must comply to your "dagster" expectations meaning - I'm
getting the below which as is, would be an issue unless I "fixed" and
ran all jobs via Dagster

A dbt project could potentially have many jobs (for testing, managed by
Dagster, manual use, etc). Rather than trying to manage all of this,
only manage the jobs that are explicitly labeled by the user.

Here, we use the presence of the environment variable cache to determine
whether or not we should populate its cache.

### How I Tested These Changes
pytest, local