Make sure we always wait for all grpc servers to spin down in tests
Summary:
Some of the grpc flakiness we are seeing on upgrade seems to stem from lingering grpc subprocesses still running under test. We previously only spun these down for launched runs due to DB contention - this PR takes a step beyond that and ensures that any grpc server process that is opened during a test is closed before that test ends.