[telemetry] fix setting __TELEMETRY_ENABLED__ flag on Dagit (#12092)
## Summary

Noticed while trying to add another python-populated Dagit flag.

The `"__TELEMETRY_ENABLED__"` placeholder string was being set to a
string value `"True"` or `"False`", which the JS does not convert into a
boolean value. This PR fixes the templating by replacing the entire
placeholder string with a literal
`true` or `false `JS bool.