[dagit] Add tests for unpartitioned case, fix for console.error in partition health (#12085)
### Summary & Motivation

This PR addresses the error surfaced in
https://elementl-workspace.slack.com/archives/C03KE7K92TT/p1675447983332539
- retrieving the partition health of an unpartitioned asset would throw
an uncaught exception because the health data suggested a one
dimensional asset but zero dimensions were present.

This PR also adds tests for a multi-dimensional asset where both
partition ranges are static, since the code path for that is slightly
different. This uncovered another small bug in the business logic [!]
where the UI would show a "solid green bar" even if a second-dimension
key was only materialized for all 1st-dimension keys with any
materialization, not all 1st dimension keys total.

### How I Tested These Changes

I added a failing test case for the datadog case and verified it's fixed
by the change. I /think/ with these updates the test coverage for this
is fairly complete.

Co-authored-by: bengotow <bgotow@elementl.com>