Test with a 64 bit after_cursor (#12093)
We have some dagster cloud tests that run this same test suite, but
against a fork of event logs that uses a bigint instead of an int.

To flex that behavior, we first manually increment our event log id
sequence beyond 32 bits. This test fails because under those conditions,
there actually are events with ids above 9999999.

Introducing a breaking migration to use bigints in dagster's event log
is well beyond the scope of this change. But incrementing this test
value to something much much larger maintains the spirit of the test
while accomodating our dagster cloud test suite.