[typing/runtime] [gql] eliminate resolver **kwargs (3) (#12242)
### Summary & Motivation

Replace resolver `**kwargs` with typed keyword args. Part of a
collection of PRs that eliminates resolver `**kwargs` across all
resolvers.

**NOTE**: The GQL schema in this PR is changed, but its a non-functional
change-- the change arises from the parameter order of the signature of
`resolve_isPipelineConfigValid`-- keyword args with defaults need to
appear at the end of the param list in a python function. This changes
the order in which the args are listed in the generated schema, but this
doesn't matter because args are always passed by name in GQL.

### How I Tested These Changes

BK. Also, correctness of added args was confirmed by manually inspecting
the field definition in the parent object.