[typing/runtime] [gql] eliminate resolver **kwargs (1) (#11724)
### Summary & Motivation

Many resolver methods currently use `**kwargs` for their arguments. This
makes for less readable and poorly typed code. This PR updates all
resolvers using `**kwargs` to instead explicitly list their arguments.
This requires various changes to the corresponding function bodies (to
access the params directly instead of `kwargs`), but the intent here is
a pure refactor with no behavior change. This also removes extraneous
`**_kwargs` in mutation signatures.

I determined how to change the signatures by looking at the
corresponding field definition, which defines the arguments. Note that
in a few cases the existing resolver body contained extraneous logic
given the type of the arg, which I removed.

### How I Tested These Changes

BK

The correctness of the signatures is tested by the lack of `**kwargs`--
if the resolver is exercised in our runtime tests, then any arguments
will be passed as keyword arguments, which will generate a runtime error
if it a corresponding parameter is not present in the signature.