Move repository_definition into subpackage (#12318)
### Summary & Motivation

I was looking into what it would take to add a pending job definition to
the codebase and ran into our repository definition code. It has
definitely sprawled and could use some modularization. In particular the
code to build the repository data class is quite the mess. My goal here
is to break up this single file into a subpackage and, in particular,
get the code to build the `CachingRepositoryData` into its own file to
we can refactor that.

This PR just moves the code into the `__init__.py` of an equivalently
named subpackage. This isolates the import changes necessary to make
this work.

### How I Tested These Changes

BK