Fix multipartitions run length encoding error (#12329)
Seen here:
https://app.datadoghq.com/apm/trace/3303427780901995569?spanID=13139862780265732778&spanViewType=errors

This PR is mostly just snapshot updates and is hopefully fairly fast to
review!

The root cause is that the first dimension has no partitions. This can
happen on an empty static partitions definition, or a time window
partitions definition where the start time is in the future.

This PR adds a check for this case, exiting early if the primary
dimension has no partition keys. The run length encoding will still
succeed if the secondary dimension has no partition keys.