refactor(databricks): divest from databricks_api in favor of databricks-cli (#12153)
### Summary & Motivation
Resolves https://github.com/dagster-io/dagster/issues/9871.

The layer of indirection provided by `databricks_api` is not needed:
begin to divest from it, in favor of `databricks-cli`, which is the
recommended approach to accessing the Databricks Python API.

A simple way to support arbitrary use cases with the
`dagster-databricks` integration is just to give a very simple way to
access the API client using our resources abstraction. This does just
that.

### How I Tested These Changes
bk, vercel docs