guide to how assets relate to ops and graphs (#12204)
### Summary & Motivation

New users often express confusion about how assets and ops relate and
when to use one vs. the other. There's arguably more to say on this
topic, but I wanted to do my best to lay out the basics in the
somewhat-limited time I have to work on this.

I don't think I have the bandwidth to do any big structural revisions,
but of course open to any and all feedback on how to polish it.

This meant to be kind of an updated version of the [Enriching with
Software-Defined
Assets](https://docs.dagster.io/guides/dagster/enriching-with-software-defined-assets#upgrading-to-software-defined-assets)
page, but targeted at new users instead of users who are using Dagster
with the pre-asset APIs.

Preview:
https://dagster-git-assets-relate-to-ops-elementl.vercel.app/guides/dagster/how-assets-relate-to-ops-and-graphs

### How I Tested These Changes

---------

Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>