[Docs RFC] Dynamic Partitions (#12227)
Adds `DynamicPartitionsDefinition` to API docs and adds a docs snippet
describing a use case for dynamic partitions.

One important thing to note is that this feature is experimental and
subject to change. Future changes may result in losing partition history
and errors of any sort. I attempted to communicate this through the info
block, curious if there are better ways of making this information very
clear.

Also not 100% sure we should make this part of docs, happy to move this
to a github discussion if we feel that will better communicate the
experimental nature of this feature.

---------

Co-authored-by: Sandy Ryza <sandy@elementl.com>
Co-authored-by: Erin Cochran <erin.k.cochran@gmail.com>