Auto infer multipartition <-> single dimension mapping (#12400)
The existing `SingleDimensionDependencyMapping` class handles the
upstream single-dimension -> downstream multipartitioned dependency
relationship. This PR refactors this class to also handle the opposite
relationship (upstream multipartition -> downstream single dimension),
renaming to `MultiToSingleDimensionMapping`.

This class previously required a `partition_dimension_name` argument,
which this PR makes optional. This argument is required when the
multipartitions definition has dimension partition defs that are
equivalent, in which case the class throws an error.

The final change is that we can now auto-infer when we should use this
partition mapping--when a single dimension partitions def matches a
dimension of a multipartitions def, and the two assets have a dependency
relationship. This PR adds tests to ensure that we choose this partition
mapping in those cases, otherwise relying on the
`IdentityPartitionMapping` or `AllPartitionMapping`.