More helpful asset key mismatch errors (#12008)
## Summary

A little QoL idea that came out of our onboarding breakout group last
week. Adds a suggestion to the error message when an upstream asset dep
is not found, listing any assets with similar keys:

```python
dagster._core.errors.DagsterInvalidDefinitionError: Input asset '["my", "asset1"]' for asset '["asset3"]' 
is not produced by any of the provided asset ops and is not one of the provided sources. 
Did you mean one of the following?
    ["my", "prefix", "asset1"]
```

Right now, uses three heuristics for finding potential candiates:
- Same asset name, similar prefix
- Similar asset name, same prefix
- Same asset name, prefix is off-by-one component (e.g. `["snowflake',
"elementl", "prod", "my_asset"] vs ["snowflake", "elementl",
"my_asset"]`

Currently the candidates are unordered and there's no limit to the
number shown because this is a quick impl/rfc. Can flesh it out if we
find it interesting.

## Test Plan

Added a few unit tests to showcase behavior.