[graphql] launch backfills over assets with different partitionings, if all roots have same partitioning (#11827)
### Summary & Motivation

Recent changes enabled the backfill engine to include assets with
different PartitionsDefinitions. However, there's still no way to launch
such a backfill. Based on some user discussions, we believe that the
main scenario for launching this kind of value involves a single root
asset and a set of assets downstream of it. I.e. "backfill X partitions
of asset A and all the partitions of downstream assets B and C that are
downstream of X".

This changes the launch backfill mutation to support asset selections
that include assets with different partitionings, as long all of the
assets are downstream of a set of root assets that are of the backfill
and have the same PartitionsDefinition. In that case, the
`partition_names` will apply to the root assets, and partitions will be
targeted for downstream assets if they descend from any of the
partitions targeted by the root assets.

### How I Tested These Changes