Support env valueFrom in Helm chart (#12425)
For https://github.com/dagster-io/dagster/issues/12325

The kubernetes spec for env vars is

```
    env:
    - name: DEMO_GREETING
      value: "Hello from the environment"
    - name: 'DD_ENTITY_ID',
       valueFrom:
         fieldRef:
           apiVersion: v1
           fieldPath: metadata.uid
  - name: S3_BUCKET
    valueFrom:
      configMapKeyRef:
        name: s3-settings
        key: bucket
```

we decided to simplify our helm chart to

```
env:
  ENV_ONE: one
  ENV_TWO: two
```

so we can't support the fancy `valueFrom`s.

**OLD**
~~This PR adds a new `envList` field for the rare users who need
`valueFrom`~~. We could also try and fit this to our existing env field,
but I think that's unnecessarily complicated. It'd need to accept
Union[Dict,List] and fork based on the type: for lists we currently make
a configmap instead of putting the env vars directly in the spec.
`valueFrom` vars would need to be put directly on the spec to support
the `fieldPath` args.

**NEW**
This PR makes `env` take either a list (the new format) or a dict (the
old). If it's a dict, we add the env vars to the configmap like we have
been. If it's a list, we add it directly to the container spec. This way
we keep backcompat for current users with `env`, but we also support the
full k8s schema for the field.