remove timestamp comparisons of code location entries to reduce OSS dagit replica spam (#12407)
### Summary & Motivation
Users running multiple replicas of dagit are getting annoyed with toasts
popping up all the time, because OSS repos are marked with the timestamp
at which dagit was loaded. Depending on which dagit instance is hit on
page load, we may end up with multiple `Definitions updated` toasts per
page load.

This isn't an issue with Cloud, because the timestamps are loaded from
the DB and are stable across dagit instances.

The solution here is to ignore timestamps and rely on id-based entries.
This should allow Cloud to override the id field for the status entries
to add in the location timestamp.

This should resolve https://github.com/dagster-io/dagster/issues/11648

### How I Tested These Changes
Setup an nginx instance, with multiple upstream dagit servers running on
different hosts. Confirmed that the toasts did not show up on page
loads.