[pythonic resources] Clean up initialization of env vars, treat resource objects as immutable (#12445)
## Summary

Restructures the way that we initialize `ConfigurableResource` objects
at runtime. Now, they are immutable, and we construct new versions of
the resource when we resolve environment variables and initialize
resources.

This makes the code cleaner and safer, and ensures that we handle
`EnvVar` resolution properly, including in nested data structures
(lists, dicts, config objects). Previously, we'd evaluate `EnvVar`s at
load time, meaning we'd error on loading the code location. This was a
bad experience and differed from current behavior.

## Test Plan

Add a whole bunch of unit tests to ensure env vars work in various
places & are only checked at runtime.