[dagit] Expose range-based asset health, use it for partition status rendering (#12302)
### Summary & Motivation

Earlier this month we started sending asset partition health to Dagit as
"ranges" of partition keys. The goal of this was to more concisely
express the typical case, in which large bands of a potentially enormous
range are all in the same state.

This PR updates the partition health bars in Dagit to take advantage of
this format, which more directly maps onto what we render (green / gray
bars). This PR finally de-couples the performance of the partition
health bars from the number of keys in the asset partition space.

This code is complicated because Dagit allows you to:

1) view partition health on both dimensions of a multi-partitioned
asset, even though the data sent to Dagit is always ranges on one
dimension with subranges on the other. This requires that dagit can
effectively "pivot" ranges onto the other axis.

2) view partition health filtered to a selection of keys. ("Eg: If I
look at just "Prod" on dimension 2, do I have all of 2022?" on dimension
1?) This is a super cool interaction in the backfill modal but requires
that Dagit can slice and subset range-based data.

3) view the aggregate health of several assets at once, flattening their
individual health ranges into a single bar that shows "partial" where
some of the assets are unmaterialized. (in the backfill modal if you
have multiple assets selected). This requires that we can "union"
overlapping ranges and generate a new summary set.


Next Steps:

### How I Tested These Changes

- New tests covering all the utility methods with cases written to
exercise all the branches in the logic.

- New tests covering the output of the range accessor on the health data
object. I wrote these tests directly from the example data and then
verified the output against the function (which I think is important
because it'd be easy to copy-paste the output into the toEqual and not
catch a bug in this...)

- I screenshotted the storybook coverage of PartitionHealthSummary and
verified the rendering is the same for a variety of health bars. This
was super useful in catching a bunch of bugs as I was writing the code.

### Follow-ups:

- There are only two callsites still using `stateForSingleDimension`.
Most importantly, the AssetPartitions page uses this and introduces an
O(length of partitions array) to implement the status dots next to the
partition rows and the filtering for "Success / Missing" on that page.
It'd be great to get rid of `stateForSingleDimension`, because it's
largely just a way-less-performant version of
`rangesForSingleDimension`.

- The PartitionState enum should really be called PartitionStatus to
match RunStatus. Will fix this on a quiet Friday at some point soon.

---------

Co-authored-by: bengotow <bgotow@elementl.com>