[dagit] Use range-based asset health for asset partitions / job partitions pages (#12434)
### Summary & Motivation

This PR moves the remaining bits of Dagit to use range-based asset
health data in the most direct form possible.

This is likely to be a big performance win, especially for the asset job
partitions page which loaded and very-inefficiently-queried the health
data of all the assets in the job. ( It was using a translation layer of
sorts mapping the ranges to a structure that was then queried
per-partition to count partitions, etc. )

One caveat of this move to range-based health is that we had to assemble
a utility library for slicing / subsetting ranges, counting keys in a
set of ranges with a given status, etc. This PR adds a few more
(hopefully the last of them), and I'm writing exhaustive test coverage
for them because the logic can be complex.

I think the range-based implementation is much more efficient, but it's
harder to grok two dimensional ranges windowed to a two dimensional
selection than it was to understand + visualize a 2D matrix of booleans.
I think there's a chance this could get /real/ complicated if/when we
layer "stale" into asset health, but at least we'll have strong test
coverage.

### How I Tested These Changes

15 new tests and a bunch of updated ones! The tests and storybooks share
mocks. This PR adds:

- Storybooks for AssetPartitions

- Basic tests that the AssetPartitions page renders and interactions
work:
+ Slicing the partition space via the top text input when a time
partition is present
+ Filtering the partitions based on status and seeing the counts / list
change
   + Clicking partitions to set focus

- Exhaustive tests for the new utility methods

---------

Co-authored-by: bengotow <bgotow@elementl.com>