Enable internal testing for writing asset cached status data (#12497)
Previously, `GraphQLEventLogStorage` could not write asset status cached
data to the asset key table. This was because only Dagit was reading and
writing these values.

Now that freshness policies also use this endpoint,
`GraphQLEventLogStorage` must be able to write cached data to the asset
key table. This PR enables testing in this case, and adds another small
test to confirm expected behavior.