[dagit] Remove unnecessary usage of <TestProvider> (#12519)
### Summary & Motivation

I went through the tests that used `<TestProvider>` and refactored or
removed it in places where it seemed straightforward. It's still in use
in 15 files, but those pass mocks or other props and probably need more
careful removal.

### How I Tested These Changes

Tests still pass! I also altered the ReexcutionDialog test "failure"
case to be a bit more interesting since we have more control over the
mocks with the MockedProvider.

Co-authored-by: bengotow <bgotow@elementl.com>