Fix multipartitions w/ fs_io_manager on windows (#12414)
Fixes an error where multipartitions cannot be used with the
fs_io_manager on windows. This occurs because windows cannot store files
containing the pipe (`|`) character which are used in multipartition
keys:
- https://github.com/dagster-io/dagster/issues/11865
- https://github.com/dagster-io/dagster/issues/12236 (duplicate)

This PR resolves this by storing each dimension in its own directory
level following the alphabetical order of the dimension names. For
backcompatibility, if attempting to read from the new file directory
format results in a file not found error, attempts to read from the old
file directory format.