Fetch multipartition key from context methods (#12512)
Reported by a user here:
https://dagster.slack.com/archives/C01U954MEER/p1677064207587029

In Dagit, we enabled launching a range of partitions as a single run.
This PR updates context methods to handle partition keys correctly
following this:
- `context.partition_key` would through an "unpartitioned run" error in
this case. This PR updates the error message to be more informative when
a partition range is provided
- partition keys fetched from the `ASSET_PARTITION_RANGE_START` and
`ASSET_PARTITION_RANGE_END` tags were the formatted complete
multipartition tag (with the "|"). But these values were returned as
strings and not as `MultiPartitionKeys`, so users were unable to use
`key.keys_by_dimension`. This PR fixes this behavior, by fetching the
partitions definition to create a `MultiPartitionKey` instead.