[graphql] empty numPartitions and partitionNames for cross-partitioning backfills (#12462)
### Summary & Motivation

This PR makes a handful of backfill fields optional, because things like
`numPartitions` and `partitionNames` are not always available for pure
asset backfills with partition mapping. The number of partitions is
/usually/ available (unless we were unable to determine the
relationships between the partition mappings), but the partition names
are never available.

@bengotow UI Changes: On the front-end this required making a few things
in the UI optional - most notably you cannot click the partition count
badge to see the partitions requested, and the run status column is
blank.

### How I Tested These Changes

To make sure this functionality doesn't get broken in the future and
ensure we preview what these cases look like when changing the backfill
table, I added two new cases to the storybook. (Bottom two in
screenshot!)

<img width="1726" alt="image"
src="https://user-images.githubusercontent.com/1037212/220788087-87db5dc5-4792-4053-bcce-b96205c8840d.png">

---------

Co-authored-by: bengotow <bgotow@elementl.com>