[dagit] Use last path component of linked assets on the asset graph + add tooltips (#12590)
This should resolve #12028

### Summary & Motivation

- Asset links now have tooltips on hover

- Asset links now show only the last path component of the asset key,
which matches the behavior of full-blown asset nodes. Given {key: ['a',
'b']} the graph just shows "b". This should make the truncation more
acceptable.

- We previously showed "abcde..." instead of "abc...gi" if there were 6
or less characters of space available. We now do this for 10 or less
characters so the asset links do prefix truncation nicely (and in
general, it seemed like "abc...gi" was not great)

This achieves the behavior Sandy mentioned in #12028.

### How I Tested These Changes

- Asset links now have storybooks!

![image](https://user-images.githubusercontent.com/1037212/221935839-382cadd3-c842-4550-9be4-44c855e52854.png)

Co-authored-by: bengotow <bgotow@elementl.com>